requirements.txt
manim
langchain-groq
python-dotenv
gtts
pydub
edge-tts asyncio
mutagen edge-tts